MCP工具模块化拆分 + config.py版本号升级至2.6
- mcp_server.py 将 _register_tools() 委托到 mcp_tools/ 各子模块
- 新增 config_tools.py, webdav_tools.py, ftp_tools.py, smb_tools.py
- 更新 _state.py 增加 FTP_SVC 引用
- config.py SOFTWARE_VERSION = "2.6"
- 所有23个测试通过
