+++ config.py
@@ -4,7 +4,7 @@
 
 SOFTWARE_NAME = "PersonalDAV"
 SOFTWARE_DESCRIPTION = "全能 DAV 服务 (CardDAV + CalDAV + WebDAV)"
-SOFTWARE_VERSION = "2.5"
+SOFTWARE_VERSION = "2.6"
 SOFTWARE_AUTHOR = "hunyanjie"
 
 # 数据库路径
