+++ mcp_tools/__init__.py
@@ -1,0 +1,1 @@
+# MCP 工具模块 — 按类别分文件
